requirements.txt
pefile
python-magic
yara-python
jinja2
